feat: M13-M17 + N95 deploy hardening (v0.1.0)
Per-usage Ollama URLs + Test buttons + worker AI hot-reload (M13 #369/#371/#372),
drag-and-drop document upload (M14 #370), settings/service-restart activity events
(M15 #373), idle-Ollama container autostop (M16 #374), and pluggable OCR engines:
RapidOCR (ONNX/OpenVINO, ~20x on N95) + device-aware recommendation + in-UI engine
selection (M17 #375).

Also N95 deploy/OOM hardening surfaced by the on-box test: worker DOKTOK_TENANT_TOKENS,
files-volume init, pgBackRest stanza/config, OCR mkldnn-off flag, rebalanced memory
caps; OCR docs (ADR-0010/0021 + ops guides) and version bump to 0.1.0.

+++ providers/openai/doktok_provider_openai/client.py
@@ -43,7 +43,7 @@ class OpenAiServerError(OpenAiError):
 
 
 def _backoff(attempt: int) -> float:
-    return _BACKOFF_BASE * (2**attempt) + random.uniform(0, 0.25)
+    return float(_BACKOFF_BASE * (2**attempt) + random.uniform(0, 0.25))
 
 
 def _retry_after(response: httpx.Response) -> float | None:
